Proposal Python bindings of SetCoverGenerator
