+++ python/src/ecole/core/core.hpp
@@ -14,6 +14,10 @@ namespace data {
 void bind_submodule(pybind11::module_ const& m);
 }
 
+namespace instance {
+void bind_submodule(pybind11::module const& m);
+}
+
 namespace observation {
 void bind_submodule(pybind11::module_ const& m);
 }
